[rigging] Split filesystem.py into a package
Break the ~1860-line rigging/filesystem.py into rigging/filesystem/ with one
module per concern:

  storage_path   StoragePath + its verbs, prefix_join, rebase_file_path, split_gcs_path
  cluster_config DataConfig, region/prefix resolution, temp storage, GCS utils
  cross_region   TransferBudget, CrossRegionGuardedFS, record_transfer
  factory        guarded url_to_fs/open_url/filesystem, atomic_rename
  mirror         MirrorFileSystem

MirrorFileSystem is the only piece that imports rigging.distributed_lock. To
keep distributed_lock out from under the package — so it can depend on
StoragePath without an import cycle — __init__ registers the mirror:// protocol
lazily by class path rather than importing mirror eagerly. Importing
rigging.filesystem no longer pulls in mirror or distributed_lock, and fsspec
imports the class on demand the first time a mirror:// filesystem is built. This
is the structural precondition #6995 asks for; migrating FsspecLease onto
StoragePath (removing the raw-fsspec exemption from #6994) is now unblocked and
left as the follow-up.

StoragePath is both a value type parsed by the low-level config/cross-region
code and a facade whose verbs resolve through the guarded factory, so
storage_path and factory would import each other. storage_path stays a leaf:
its verbs import url_to_fs/open_url at call time, so nothing else in the package
(nor distributed_lock) sits above it.

__init__ re-exports the public API so the ~200 `from rigging.filesystem import X`
sites (137 of them importing StoragePath) keep working. The cluster-config module
is named cluster_config, not data_config, because a submodule named data_config
would shadow the re-exported data_config() function.

Tests are split to mirror the layout (test_storage_path/test_cluster_config/
test_region_routing/test_cross_region/test_factory/test_mirror). Monkeypatch/patch
sites that reach module internals (urllib, marin_region, get_bucket_location) now
target the owning submodule.

Fixes #6995
