+++ lib/rigging/tests/test_region_routing.py
@@ -1,32 +1,24 @@
 # Copyright The Marin Authors
 # SPDX-License-Identifier: Apache-2.0
 
-"""Tests for region resolution, temp-bucket routing, cross-region guards, and the
-guarded filesystem entry points. See test_data_config.py for DataConfig loading and
-resolved_root(), and test_filesystem.py for StoragePath."""
+"""Tests for region resolution and temp-bucket routing (the cluster_config region
+aspect). See test_cluster_config.py for DataConfig loading and resolved_root(),
+test_cross_region.py for the transfer budget and guard, and test_factory.py for the
+guarded entry points."""
 
-import concurrent.futures
 import os
 from dataclasses import dataclass
 from pathlib import Path
 from unittest.mock import MagicMock, patch
 
 import pytest
-from rigging.filesystem import (
-    MARIN_CROSS_REGION_OVERRIDE_ENV,
-    CrossRegionGuardedFS,
-    TransferBudget,
-    TransferBudgetExceeded,
-    _regions_match,
+from rigging.filesystem.cluster_config import (
     check_gcs_paths_same_region,
     collect_gcs_paths,
-    filesystem,
     marin_region,
     marin_temp_bucket,
-    open_url,
     region_from_metadata,
     region_from_prefix,
-    url_to_fs,
 )
 
 
@@ -40,14 +32,14 @@ def _mock_urlopen(zone_bytes: bytes) -> MagicMock:
 
 def test_region_from_metadata_parses_zone():
     with patch(
-        "rigging.filesystem.urllib.request.urlopen",
+        "rigging.filesystem.cluster_config.urllib.request.urlopen",
         return_value=_mock_urlopen(b"projects/12345/zones/us-central2-b"),
     ):
         assert region_from_metadata() == "us-central2"
 
 
 def test_region_from_metadata_returns_none_on_failure():
-    with patch("rigging.filesystem.urllib.request.urlopen", side_effect=OSError("not on GCP")):
+    with patch("rigging.filesystem.cluster_config.urllib.request.urlopen", side_effect=OSError("not on GCP")):
         assert region_from_metadata() is None
 
 
@@ -68,14 +60,15 @@ def test_region_from_prefix(prefix, expected):
 
 def test_marin_region_from_metadata():
     with patch(
-        "rigging.filesystem.urllib.request.urlopen", return_value=_mock_urlopen(b"projects/12345/zones/us-east1-c")
+        "rigging.filesystem.cluster_config.urllib.request.urlopen",
+        return_value=_mock_urlopen(b"projects/12345/zones/us-east1-c"),
     ):
         assert marin_region() == "us-east1"
 
 
 def test_marin_region_from_env_prefix():
     with (
-        patch("rigging.filesystem.urllib.request.urlopen", side_effect=OSError("not on GCP")),
+        patch("rigging.filesystem.cluster_config.urllib.request.urlopen", side_effect=OSError("not on GCP")),
         patch.dict(os.environ, {"MARIN_PREFIX": "gs://marin-us-west4/scratch"}),
     ):
         assert marin_region() == "us-west4"
@@ -84,7 +77,7 @@ def test_marin_region_from_env_prefix():
 def test_marin_region_normalizes_eu_west4():
     """Regression test: marin-eu-west4 bucket must resolve to europe-west4."""
     with (
-        patch("rigging.filesystem.urllib.request.urlopen", side_effect=OSError("not on GCP")),
+        patch("rigging.filesystem.cluster_config.urllib.request.urlopen", side_effect=OSError("not on GCP")),
         patch.dict(os.environ, {"MARIN_PREFIX": "gs://marin-eu-west4/tokenized"}),
     ):
         assert marin_region() == "europe-west4"
@@ -92,7 +85,7 @@ def test_marin_region_normalizes_eu_west4():
 
 def test_marin_region_none_when_unresolvable():
     with (
-        patch("rigging.filesystem.urllib.request.urlopen", side_effect=OSError("not on GCP")),
+        patch("rigging.filesystem.cluster_config.urllib.request.urlopen", side_effect=OSError("not on GCP")),
         patch.dict(os.environ, {}, clear=True),
     ):
         assert marin_region() is None
@@ -100,7 +93,7 @@ def test_marin_region_none_when_unresolvable():
 
 def test_marin_temp_bucket_from_metadata():
     with patch(
-        "rigging.filesystem.urllib.request.urlopen",
+        "rigging.filesystem.cluster_config.urllib.request.urlopen",
         return_value=_mock_urlopen(b"projects/12345/zones/us-central2-b"),
     ):
         assert marin_temp_bucket(ttl_days=30, prefix="compilation-cache") == (
@@ -110,7 +103,7 @@ def test_marin_temp_bucket_from_metadata():
 
 def test_marin_temp_bucket_from_env_prefix():
     with (
-        patch("rigging.filesystem.urllib.request.urlopen", side_effect=OSError("not on GCP")),
+        patch("rigging.filesystem.cluster_config.urllib.request.urlopen", side_effect=OSError("not on GCP")),
         patch.dict(os.environ, {"MARIN_PREFIX": "gs://marin-us-east1/scratch"}),
     ):
         assert marin_temp_bucket(ttl_days=3, prefix="zephyr") == "gs://marin-us-east1/tmp/ttl=3d/zephyr"
@@ -119,7 +112,7 @@ def test_marin_temp_bucket_from_env_prefix():
 def test_marin_temp_bucket_eu_west4_uses_main_bucket_alias():
     """eu-west4 region resolves to the canonical europe-west4 marin-eu-west4 bucket."""
     with (
-        patch("rigging.filesystem.urllib.request.urlopen", side_effect=OSError("not on GCP")),
+        patch("rigging.filesystem.cluster_config.urllib.request.urlopen", side_effect=OSError("not on GCP")),
         patch.dict(os.environ, {"MARIN_PREFIX": "gs://marin-eu-west4/scratch"}),
     ):
         assert marin_temp_bucket(ttl_days=1, prefix="ferry") == "gs://marin-eu-west4/tmp/ttl=1d/ferry"
@@ -127,7 +120,7 @@ def test_marin_temp_bucket_eu_west4_uses_main_bucket_alias():
 
 def test_marin_temp_bucket_uses_source_prefix_region():
     with (
-        patch("rigging.filesystem.urllib.request.urlopen", side_effect=OSError("not on GCP")),
+        patch("rigging.filesystem.cluster_config.urllib.request.urlopen", side_effect=OSError("not on GCP")),
         patch.dict(os.environ, {"MARIN_PREFIX": "gs://marin-us-central1/scratch"}),
     ):
         assert marin_temp_bucket(
@@ -139,7 +132,7 @@ def test_marin_temp_bucket_uses_source_prefix_region():
 
 def test_marin_temp_bucket_uses_source_prefix_region_from_local_launcher():
     with (
-        patch("rigging.filesystem.urllib.request.urlopen", side_effect=OSError("not on GCP")),
+        patch("rigging.filesystem.cluster_config.urllib.request.urlopen", side_effect=OSError("not on GCP")),
         patch.dict(os.environ, {}, clear=True),
     ):
         assert marin_temp_bucket(
@@ -152,7 +145,7 @@ def test_marin_temp_bucket_uses_source_prefix_region_from_local_launcher():
 def test_marin_temp_bucket_r2_uses_bucket_root_ttl_path():
     """An R2 marin prefix resolves temp to the bucket root, dropping the marin/ subdir."""
     with (
-        patch("rigging.filesystem.urllib.request.urlopen", side_effect=OSError("not on GCP")),
+        patch("rigging.filesystem.cluster_config.urllib.request.urlopen", side_effect=OSError("not on GCP")),
         patch.dict(os.environ, {"MARIN_PREFIX": "s3://marin-na/marin"}),
     ):
         assert marin_temp_bucket(ttl_days=1, prefix="zephyr") == "s3://marin-na/tmp/ttl=1d/zephyr"
@@ -161,7 +154,7 @@ def test_marin_temp_bucket_r2_uses_bucket_root_ttl_path():
 
 def test_marin_temp_bucket_r2_uses_source_prefix_bucket():
     with (
-        patch("rigging.filesystem.urllib.request.urlopen", side_effect=OSError("not on GCP")),
+        patch("rigging.filesystem.cluster_config.urllib.request.urlopen", side_effect=OSError("not on GCP")),
         patch.dict(os.environ, {}, clear=True),
     ):
         assert (
@@ -174,7 +167,7 @@ def test_marin_temp_bucket_r2_source_prefix_overrides_gcs_launcher():
     """An explicit R2 source_prefix wins over a gs:// MARIN_PREFIX and GCP metadata."""
     with (
         patch(
-            "rigging.filesystem.urllib.request.urlopen",
+            "rigging.filesystem.cluster_config.urllib.request.urlopen",
             return_value=_mock_urlopen(b"projects/12345/zones/us-central1-a"),
         ),
         patch.dict(os.environ, {"MARIN_PREFIX": "gs://marin-us-central1/scratch"}),
@@ -188,7 +181,7 @@ def test_marin_temp_bucket_r2_source_prefix_overrides_gcs_launcher():
 def test_marin_temp_bucket_unknown_s3_bucket_falls_back_to_flat_path():
     """Unknown S3 buckets have no lifecycle rules, so they get the flat non-TTL fallback."""
     with (
-        patch("rigging.filesystem.urllib.request.urlopen", side_effect=OSError("not on GCP")),
+        patch("rigging.filesystem.cluster_config.urllib.request.urlopen", side_effect=OSError("not on GCP")),
         patch.dict(os.environ, {"MARIN_PREFIX": "s3://some-other-bucket/marin"}),
     ):
         assert marin_temp_bucket(ttl_days=1, prefix="x") == "s3://some-other-bucket/marin/tmp/x"
@@ -197,7 +190,7 @@ def test_marin_temp_bucket_unknown_s3_bucket_falls_back_to_flat_path():
 def test_marin_temp_bucket_falls_back_to_marin_prefix_when_no_region():
     # Unknown region in MARIN_PREFIX → no entry in DataConfig.region_buckets → falls back to marin_prefix/tmp
     with (
-        patch("rigging.filesystem.urllib.request.urlopen", side_effect=OSError("not on GCP")),
+        patch("rigging.filesystem.cluster_config.urllib.request.urlopen", side_effect=OSError("not on GCP")),
         patch.dict(os.environ, {"MARIN_PREFIX": "gs://marin-antarctica-south1/scratch"}),
     ):
         result = marin_temp_bucket(ttl_days=30)
@@ -206,7 +199,7 @@ def test_marin_temp_bucket_falls_back_to_marin_prefix_when_no_region():
 
 def test_marin_temp_bucket_local_fallback_when_unresolvable():
     with (
-        patch("rigging.filesystem.urllib.request.urlopen", side_effect=OSError("not on GCP")),
+        patch("rigging.filesystem.cluster_config.urllib.request.urlopen", side_effect=OSError("not on GCP")),
         patch.dict(os.environ, {}, clear=True),
     ):
         assert marin_temp_bucket(ttl_days=30, prefix="iris-logs") == "file:///tmp/marin/tmp/iris-logs"
@@ -214,14 +207,15 @@ def test_marin_temp_bucket_local_fallback_when_unresolvable():
 
 def test_marin_temp_bucket_no_prefix():
     with patch(
-        "rigging.filesystem.urllib.request.urlopen", return_value=_mock_urlopen(b"projects/12345/zones/us-east1-c")
+        "rigging.filesystem.cluster_config.urllib.request.urlopen",
+        return_value=_mock_urlopen(b"projects/12345/zones/us-east1-c"),
     ):
         assert marin_temp_bucket(ttl_days=14) == "gs://marin-us-east1/tmp/ttl=14d"
 
 
 def test_marin_temp_bucket_strips_prefix_slashes():
     with patch(
-        "rigging.filesystem.urllib.request.urlopen",
+        "rigging.filesystem.cluster_config.urllib.request.urlopen",
         return_value=_mock_urlopen(b"projects/12345/zones/us-central1-a"),
     ):
         assert marin_temp_bucket(ttl_days=3, prefix="/foo/bar/") == "gs://marin-us-central1/tmp/ttl=3d/foo/bar"
@@ -231,10 +225,10 @@ def test_marin_temp_bucket_rounds_up_unsupported_ttl(caplog):
     """ttl_days values between configured points round up to the next one with a warning."""
     with (
         patch(
-            "rigging.filesystem.urllib.request.urlopen",
+            "rigging.filesystem.cluster_config.urllib.request.urlopen",
             return_value=_mock_urlopen(b"projects/12345/zones/us-east1-c"),
         ),
-        caplog.at_level("WARNING", logger="rigging.filesystem"),
+        caplog.at_level("WARNING", logger="rigging.filesystem.cluster_config"),
     ):
         # 10 → 14, 15 → 30
         assert marin_temp_bucket(ttl_days=10, prefix="zephyr") == "gs://marin-us-east1/tmp/ttl=14d/zephyr"
@@ -247,10 +241,10 @@ def test_marin_temp_bucket_clamps_above_max_ttl(caplog):
     """ttl_days above the configured maximum clamp to the max with a warning."""
     with (
         patch(
-            "rigging.filesystem.urllib.request.urlopen",
+            "rigging.filesystem.cluster_config.urllib.request.urlopen",
             return_value=_mock_urlopen(b"projects/12345/zones/us-east1-c"),
         ),
-        caplog.at_level("WARNING", logger="rigging.filesystem"),
+        caplog.at_level("WARNING", logger="rigging.filesystem.cluster_config"),
     ):
         assert marin_temp_bucket(ttl_days=100) == "gs://marin-us-east1/tmp/ttl=30d"
     assert any("clamping to 30" in rec.message for rec in caplog.records)
@@ -330,202 +324,3 @@ def test_collect_gcs_paths_recurses_and_skips_prefixes():
         ("config.nested.path", "gs://bucket/nested"),
         ("config.set_field[0]", "gs://bucket/from_set"),
     ]
-
-
-# ---------------------------------------------------------------------------
-# _regions_match tests
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.parametrize(
-    "vm_region, bucket_location, expected",
-    [
-        ("us-central1", "us-central1", True),
-        ("US-Central1", "us-central1", True),
-        ("europe-west4", "europe-west4", True),
-        ("us-central1", "eu-west4", False),
-        ("us-central1", "us", True),
-        ("us-east1", "us", True),
-        ("europe-west4", "eu", True),
-        ("asia-northeast1", "asia", True),
-        ("eu-west4", "us", False),
-        ("us-central1", "asia", False),
-    ],
-)
-def test_regions_match(vm_region, bucket_location, expected):
-    assert _regions_match(vm_region, bucket_location) is expected
-
-
-# ---------------------------------------------------------------------------
-# TransferBudget tests
-# ---------------------------------------------------------------------------
-
-
-def test_budget_records_and_blocks():
-    budget = TransferBudget(limit_bytes=1000)
-    budget.record(400, "a")
-    budget.record(400, "b")
-    assert budget.bytes_used == 800
-
-    with pytest.raises(TransferBudgetExceeded, match="transfer budget exceeded"):
-        budget.record(300, "c")
-
-    # Counter unchanged on failure.
-    assert budget.bytes_used == 800
-
-
-def test_budget_thread_safety():
-    budget = TransferBudget(limit_bytes=10 * 1024 * 1024)
-
-    def record_batch():
-        for _ in range(100):
-            budget.record(1, "x")
-
-    with concurrent.futures.ThreadPoolExecutor(max_workers=8) as pool:
-        list(pool.map(lambda _: record_batch(), range(8)))
-
-    assert budget.bytes_used == 800
-
-
-# ---------------------------------------------------------------------------
-# CrossRegionGuardedFS tests
-# ---------------------------------------------------------------------------
-
-
-class _FakeGCSFS:
-    protocol = "gs"
-
-    def __init__(self) -> None:
-        self._files: dict[str, bytes] = {}
-
-    def add_file(self, path: str, data: bytes) -> None:
-        self._files[path] = data
-
-    def size(self, path: str) -> int | None:
-        data = self._files.get(path)
-        return len(data) if data is not None else None
-
-    def open(self, path: str, mode: str = "rb", **kwargs):
-        return self._files.get(path, b"")
-
-    def cat_file(self, path: str, start=None, end=None, **kwargs) -> bytes:
-        return self._files.get(path, b"")
-
-    def cat(self, path, recursive=False, on_error="raise", **kwargs):
-        if isinstance(path, str):
-            return self._files.get(path, b"")
-        return {p: self._files.get(p, b"") for p in path}
-
-    def get_file(self, rpath: str, lpath: str, **kwargs) -> None:
-        pass
-
-    def get(self, rpath, lpath, recursive=False, **kwargs) -> None:
-        pass
-
-    def exists(self, path: str) -> bool:
-        return path in self._files
-
-
-@pytest.fixture()
-def budget():
-    return TransferBudget(limit_bytes=1024)
-
-
-def test_guarded_fs_charges_budget_for_cross_region_reads(budget):
-    fs = _FakeGCSFS()
-    for i in range(3):
-        fs.add_file(f"remote-bucket/f{i}.bin", b"x" * 400)
-
-    guarded = CrossRegionGuardedFS(fs, cross_region_checker=lambda _: True, budget=budget)
-
-    guarded.open("remote-bucket/f0.bin", "rb")
-    guarded.open("remote-bucket/f1.bin", "rb")
-    assert budget.bytes_used == 800
-
-    with pytest.raises(TransferBudgetExceeded):
-        guarded.open("remote-bucket/f2.bin", "rb")
-
-
-def test_guarded_fs_skips_same_region(budget):
-    fs = _FakeGCSFS()
-    fs.add_file("local-bucket/big.bin", b"x" * 9999)
-
-    guarded = CrossRegionGuardedFS(fs, cross_region_checker=lambda _: False, budget=budget)
-    guarded.open("local-bucket/big.bin", "rb")
-    assert budget.bytes_used == 0
-
-
-def test_guarded_fs_override_env_bypasses_budget(budget):
-    fs = _FakeGCSFS()
-    fs.add_file("remote-bucket/big.bin", b"x" * 2000)
-
-    guarded = CrossRegionGuardedFS(fs, cross_region_checker=lambda _: True, budget=budget)
-
-    with patch.dict(os.environ, {MARIN_CROSS_REGION_OVERRIDE_ENV: "testuser"}):
-        guarded.open("remote-bucket/big.bin", "rb")
-
-    assert budget.bytes_used == 0
-
-
-def test_guarded_fs_write_mode_skips_guard(budget):
-    fs = _FakeGCSFS()
-    fs.add_file("remote-bucket/big.bin", b"x" * 2000)
-
-    guarded = CrossRegionGuardedFS(fs, cross_region_checker=lambda _: True, budget=budget)
-    guarded.open("remote-bucket/big.bin", "wb")
-    assert budget.bytes_used == 0
-
-
-@pytest.mark.parametrize(
-    "method, args",
-    [
-        ("cat_file", ("remote-bucket/f.bin",)),
-        ("cat", (["remote-bucket/f.bin"],)),
-        ("get_file", ("remote-bucket/f.bin", "/tmp/local")),
-        ("get", ("remote-bucket/f.bin", "/tmp/local")),
-    ],
-    ids=["cat_file", "cat_list", "get_file", "get"],
-)
-def test_guarded_fs_all_read_methods_charge_budget(budget, method, args):
-    fs = _FakeGCSFS()
-    fs.add_file("remote-bucket/f.bin", b"x" * 100)
-
-    guarded = CrossRegionGuardedFS(fs, cross_region_checker=lambda _: True, budget=budget)
-    getattr(guarded, method)(*args)
-    assert budget.bytes_used == 100
-
-
-def test_guarded_fs_delegates_non_read_methods():
-    fs = _FakeGCSFS()
-    fs.add_file("bucket/file.txt", b"hello")
-
-    guarded = CrossRegionGuardedFS(fs, cross_region_checker=lambda _: True)
-    assert guarded.exists("bucket/file.txt") is True
-    assert guarded.exists("bucket/nope.txt") is False
-
-
-# ---------------------------------------------------------------------------
-# Guarded entry point tests
-# ---------------------------------------------------------------------------
-
-
-def test_url_to_fs_does_not_wrap_local(tmp_path):
-    test_file = tmp_path / "test.txt"
-    test_file.write_text("hello")
-
-    fs, _path = url_to_fs(str(test_file))
-    assert not isinstance(fs, CrossRegionGuardedFS)
-
-
-def test_open_url_local_file(tmp_path):
-    test_file = tmp_path / "test.txt"
-    test_file.write_text("hello")
-
-    result = open_url(str(test_file), "r")
-    with result as f:
-        assert f.read() == "hello"
-
-
-def test_filesystem_local():
-    fs = filesystem("file")
-    assert not isinstance(fs, CrossRegionGuardedFS)
