+++ tests/test_grug_launch_checkpoint_paths.py
@@ -29,7 +29,7 @@ def test_build_grug_run_config_sets_temporary_checkpoint_base_path():
     """
     output_path = "gs://marin-us-east5/experiments/grug/base-trial"
     with (
-        patch("rigging.filesystem.urllib.request.urlopen", side_effect=OSError("not on GCP")),
+        patch("rigging.filesystem.cluster_config.urllib.request.urlopen", side_effect=OSError("not on GCP")),
         patch.dict(os.environ, {"MARIN_PREFIX": "gs://marin-us-central1/scratch"}),
     ):
         run_config = build_grug_run_config(
